# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    GetTelemetryTraceResponse:
      type: object
      properties:
        trace:
          $ref: "#/components/schemas/TelemetryTraceDto"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListSlowSpansResponse:
      type: object
      properties:
        spans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
        total:
          type: integer
          format: int32
        truncated:
          type: boolean
    ListTelemetryErrorsResponse:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryErrorGroupDto"
        total:
          type: integer
          format: int32
        truncated:
          type: boolean
...
    ListTelemetrySourcesResponse:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySource"
    ListTelemetryTracesResponse:
      type: object
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryTraceSummary"
        total:
          type: integer
          format: int32
        truncated:
          type: boolean
    SearchTelemetryLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
        total:
          type: integer
          format: int32
        truncated:
          type: boolean
    SearchTelemetryRecordsRequest:
      type: object
      required:
      - subscribe
      properties:
        subscribe:
          type: array
          items: {}
          description: "The groups of a live subscribe frame, exactly as README \"\
            The live stream\" defines them: [{\"conditions\": [{\"field\", \"op\"\
            , \"value\", \"key\"?}]}]. Groups OR, conditions AND; [] matches nothing."
        since:
          type: string
          format: date-time
          description: "Inclusive lower bound, an ISO-8601 instant. Absent: no lower\
            \ bound."
        until:
          type: string
          format: date-time
          description: "Inclusive upper bound, an ISO-8601 instant. Absent: now."
        limit:
          type: integer
          format: int32
          description: "At most this many records, 1..1000. Absent: 200."
        source:
          type: string
          description: "A key from /telemetry/sources. Absent: every source."
    SearchTelemetryRecordsResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryStreamFrame"
        truncated:
          type: boolean
        bufferedSince:
          $ref: "#/components/schemas/Instant"
    SpanEvent:
      type: object
      properties:
        name:
          type: string
        epochNanos:
          type: integer
          format: int64
        attributes:
          type: object
          additionalProperties:
            type: string
        exception:
          type: boolean
    TelemetryErrorGroupDto:
      type: object
      properties:
        traceId:
          type: string
        serviceName:
          type: string
        errorSpans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
        errorLogs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
    TelemetryLogDto:
      type: object
      properties:
        epochNanos:
          type: integer
          format: int64
        severityNumber:
          type: integer
          format: int32
        severityText:
          type: string
        body:
          type: string
        traceId:
          type: string
        spanId:
          type: string
        serviceName:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
        resourceAttributes:
          type: object
          additionalProperties:
            type: string
...
    TelemetrySource:
      type: object
      description: One addressable bucket of the telemetry buffer.
      properties:
        key:
          type: string
        kind:
          $ref: "#/components/schemas/TelemetrySourceKind"
        label:
          type: string
        repositoryId:
          type: string
        workspaceId:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySourceService"
        spans:
          type: integer
          format: int32
        logs:
          type: integer
          format: int32
        metricSeries:
          type: integer
          format: int32
        bytes:
          type: integer
          format: int64
        oldestReceivedAt:
          $ref: "#/components/schemas/Instant"
        newestReceivedAt:
          $ref: "#/components/schemas/Instant"
    TelemetrySourceKind:
      type: string
      enum:
      - SERVICE
      - WORKSPACE
      - UNSCOPED
    TelemetrySourceService:
      type: object
      properties:
        name:
          type: string
        spans:
          type: integer
          format: int32
        logs:
          type: integer
          format: int32
        metricSeries:
          type: integer
          format: int32
    TelemetrySpanDto:
      type: object
      properties:
        traceId:
          type: string
        spanId:
          type: string
        parentSpanId:
          type: string
        serviceName:
          type: string
        scopeName:
          type: string
        name:
          type: string
        kind:
          type: string
        startEpochNanos:
          type: integer
          format: int64
        durationMs:
          type: integer
          format: int64
        status:
          type: string
        statusMessage:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/SpanEvent"
        resourceAttributes:
          type: object
          additionalProperties:
            type: string
    TelemetryStoreCaps:
      type: object
      properties:
        spansPerSource:
          type: integer
          format: int32
        logsPerSource:
          type: integer
          format: int32
        metricSeriesPerSource:
          type: integer
          format: int32
    TelemetryStoreState:
      type: object
      description: The in-memory buffer's own state.
      properties:
        startedAt:
          $ref: "#/components/schemas/Instant"
        totalBytes:
          type: integer
          format: int64
        maxTotalBytes:
          type: integer
          format: int64
        maxBytesPerSource:
          type: integer
          format: int64
        caps:
          $ref: "#/components/schemas/TelemetryStoreCaps"
        sourceCount:
          type: integer
          format: int32
        evictedSpans:
          type: integer
          format: int64
        evictedLogs:
          type: integer
          format: int64
        droppedMetricSeries:
          type: integer
          format: int64
    TelemetryStreamFrame:
      type: object
      properties:
        kind:
          type: string
        receivedAtMillis:
          type: integer
          format: int64
        source:
          type: string
        record: {}
    TelemetryTraceDto:
      type: object
      properties:
        traceId:
          type: string
        spans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
    TelemetryTraceSummary:
      type: object
      description: One row of the trace list.
      properties:
        traceId:
          type: string
        rootName:
          type: string
        rootService:
          type: string
        rootRoute:
          type: string
        services:
          type: array
          items:
            type: string
        startEpochNanos:
          type: integer
          format: int64
        durationMs:
          type: integer
          format: int64
        spanCount:
          type: integer
          format: int32
        errorSpanCount:
          type: integer
          format: int32
        hasException:
          type: boolean
        rootMissing:
          type: boolean
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /observability/api/telemetry/errors:
    get:
      operationId: listTelemetryErrors
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 200
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: service
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: source
        in: query
        schema:
          type: string
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTelemetryErrorsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Errors
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/logs:
    get:
      operationId: searchTelemetryLogs
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 200
      - name: minSeverity
        in: query
        schema:
          type: string
      - name: query
        in: query
        schema:
          type: string
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: service
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: source
        in: query
        schema:
          type: string
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchTelemetryLogsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Logs
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
...
  /observability/api/telemetry/records/search:
    post:
      operationId: searchTelemetryRecords
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTelemetryRecordsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchTelemetryRecordsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Search Records
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/slow-spans:
    get:
      operationId: listSlowSpans
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 200
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: service
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        schema:
          type: string
          default: duration
      - name: source
        in: query
        schema:
          type: string
      - name: thresholdMs
        in: query
        schema:
          type: integer
          format: int64
          default: 500
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSlowSpansResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Slow Spans
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/sources:
    get:
      operationId: listTelemetrySources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTelemetrySourcesResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Sources
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/store:
    get:
      operationId: getTelemetryStore
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryStoreState"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Store
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/traces:
    get:
      operationId: listTelemetryTraces
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 200
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: service
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        schema:
          type: string
          default: recent
      - name: source
        in: query
        schema:
          type: string
      - name: thresholdMs
        in: query
        schema:
          type: integer
          format: int64
          default: 0
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTelemetryTracesResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Traces
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /observability/api/telemetry/traces/{traceId}:
    get:
      operationId: getTelemetryTrace
      parameters:
      - name: traceId
        in: path
        required: true
        schema:
          type: string
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: source
        in: query
        schema:
          type: string
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTelemetryTraceResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Trace
      tags:
      - Workspace Telemetry Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
info:
  title: qits-observability API
  version: 2026.930.90310